Add files via upload
This project describes the spendings of the G20 countries in the 3 sectors viz. Education, Healthcare & Military. The data is obtained from worldbank.com. The data presented is for 5 years from 2012 to 2016, and the countries for which the data is being presented are Australia, Canada, China, Germany, United Kingdom, India, Japan, Mexico, Russia, United States & South Africa.